requirements.txt
boto3
botocore
cffi
openai
psycopg2-binary
langchain
pinecone-client
requests
tiktoken
python-dotenv
pydantic
mysql-connector-python
pymysql
pandas
qdrant-client
singlestoredb
weaviate-client
azure-storage-blob
langchain-community
google-cloud-storage
snowflake-connector-python
stripe
vecs
simple-salesforce
google-generativeai
langchain-google-genai
langchain-cohere
langchain-anthropic
anthropic
pympler
unstructured[all-docs]
dropbox
zenpy
lancedb
pyyaml
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
unstructured-client
box-sdk-gen
python-magic
